refactor(deps-restorer): split the virtual-store snapshot plan and partition
`CreateVirtualStore::run` was 794 lines, two thirds of its file, and
read as one sequence with no boundary between the decisions it makes.
Extract the two largest phases into modules, leaving `run` at 517:

- `snapshot_plan` — which snapshots this install must materialize and
  which the current-lockfile check leaves alone, plus the store-index
  cache key each is looked up by. The asymmetry is worth a name:
  survivors take the strict key derivation because a malformed
  resolution must fail before the warm batch starts, while skipped
  snapshots take a lenient pass because a per-snapshot error there
  costs only a prefetch row.
- `partition` — the warm/cold split and the store-index rows the build
  and bin phases read. Skipped snapshots are walked here too: they have
  no link work, but dropping their side-effects rows would make the
  build phase re-run approved scripts on every warm reinstall.

`SnapshotWithCacheKey` was a type alias local to `run`; it moves to
module scope so both phases can name it.

Pure moves. No behaviour change.
